requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.2
spotipy==2.25.1